Make BlkDevicesTable more generic
